Fixed duplicate code
