+++ pages/bio1.tsx
@@ -1,42 +1,22 @@
 import { useState, useEffect } from "react";
 import { Box, Card, Container, Divider, Flex, Header, Problem, Solution, SolutionSkeleton, Text, VStack } from "components";
 import { bio1Problems, pages, ProblemInfo } from "content";
-import { getProblemFromURL } from "./util";
+import { getProblemFromURL, loadProblemFromURL, setProblemWrapper } from "pages-lib/util";
 
 export default function BIO1() {
   const [problem, setProblem] = useState<ProblemInfo | null>(null);
 
-  function _setProblem(newProblem: ProblemInfo | null, pushToHistory: boolean = true) {
-    if (pushToHistory) {
-      /*
-       * Pushes problem URL to the session history
-       * Allows user to navigate between problems they've viewed
-      */
-      const url = new URL(location.toString());
-      if (newProblem !== null)
-        url.searchParams.set("problem", newProblem.display);
-      else
-        url.searchParams.delete("problem");      
-      history.pushState({}, "", url)
-    }
-    setProblem(newProblem);
-  }
-
   useEffect(() => {
-    let problemInURL = getProblemFromURL(bio1Problems);
-    console.log(problemInURL?.display);
-    if (problemInURL)
-      // If the problem parameter is already in the URL, there's no need to push it again to session history
-      _setProblem(problemInURL, false);
+    loadProblemFromURL(bio1Problems, setProblem);
     onpopstate = () => {
       /*
        * Callback when the user navigates session history.
       */
       let problemInURL = getProblemFromURL(bio1Problems);
       if (problemInURL)
-        _setProblem(problemInURL, false);
+        setProblemWrapper(problemInURL, setProblem, false);
       else
-        _setProblem(null, false);
+        setProblemWrapper(null, setProblem, false);
     }
   }, []);
 
@@ -70,7 +50,7 @@ export default function BIO1() {
                       ...bio1Problems.map((problem) => {
                         return (
                           <Box display="flex" p={1} key={problem.display}>
-                            <Problem problem={problem} onChoose={() => _setProblem(problem)} />
+                            <Problem problem={problem} onChoose={() => setProblemWrapper(problem, setProblem)} />
                           </Box>
                         );
                       })
